requirements.txt
python-dotenv~=1.0.0
paho-mqtt~=1.6.1
pymongo~=4.5.0
requests~=2.28.1
matplotlib~=3.8.0
python-dateutil~=2.8.2